+++ src/main/java/org/ssu/edu/teachua/api/models/location_city/LocationCity.java
@@ -3,11 +3,12 @@ package org.ssu.edu.teachua.api.models.location_city;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+
 @Data
 @AllArgsConstructor
 public class LocationCity {
-    private int id;
+    private Integer id;
     private String name;
-    private int latitude;
-    private int longitude;
+    private Double latitude;
+    private Double longitude;
 }
